+++ examples/selenium/mocha/lib/blazedemo.js
@@ -1,0 +1,11 @@
+exports.homePage = function() {
+    return 'http://blazedemo.com/';
+};
+
+exports.homePageTitle = function() {
+    return 'Simple Flights App'
+};
+
+exports.reserve = function() {
+    return 'reserve.php'
+};